+++ NEWS.md
@@ -1,3 +1,8 @@
+# tidyREDCap 0.2.0 
+
+* Added make_choose_one_table() function
+* Added make_instrument() function
+
 # tidyREDCap 0.1.0 (CRAN release)
 
 * Fixed title capitalization.